+++ src/modules/authorities.js
@@ -1,7 +1,7 @@
-var FS = require('./FamilySearch'),
-    utils = require('./utils'),
-    Date = require('./date'),
-    Place = require('./place');
+var FS = require('./../FamilySearch'),
+    utils = require('./../utils'),
+    Date = require('./../classes/date'),
+    Place = require('./../classes/place');
 
 /**
  * @ngdoc overview
@@ -39,9 +39,12 @@ FS.prototype.getDate = function(date, opts) {
   return self.plumbing.get(self.helpers.getAuthoritiesServerUrl('/authorities/v1/date'), params, {'Accept': 'application/json'}, opts,
     utils.compose(
       utils.objectExtender({getDate: function() { return utils.maybe(utils.maybe(this.dates).date)[0]; }}),
-      utils.constructorSetter(Date, 'date', function(response) {
-        return response.dates;
-      })
+      function(response){
+        utils.forEach(response.dates.date, function(date, index, obj){
+          obj[index] = self.createDate(date);
+        });
+        return response;
+      }
     ));
 };
 
@@ -73,9 +76,12 @@ FS.prototype.getPlaceSearch = function(place, opts) {
   return self.plumbing.get(self.helpers.getAuthoritiesServerUrl('/authorities/v1/place'), params, {'Accept': 'application/json'}, opts,
     utils.compose(
       utils.objectExtender({getPlaces: function() { return utils.maybe(this.places).place; }}),
-      utils.constructorSetter(Place, 'place', function(response) {
-        return response.places;
-      })
+      function(response){
+        utils.forEach(response.places.place, function(place, index, obj){
+          obj[index] = self.createPlace(place);
+        });
+        return response;
+      }
     ));
 };
